+++ database/migrations/2023_05_09_121358_create_paiements_table.php
@@ -15,9 +15,9 @@ class CreatePaiementsTable extends Migration
     {
         Schema::create('paiements', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('id_Intervenant');
+            $table->unsignedBigInteger('id_Intervenant');
             $table->foreign('id_Intervenant')->references('id')->on('enseignants')->onDelete('cascade');
-            $table->unsignedInteger('id_Etab');
+            $table->unsignedBigInteger('id_Etab');
             $table->foreign('id_Etab')->references('id')->on('etablissements');
             $table->float('VH')->default(0);
             $table->float('Taux_H');
